\name{GRASP-internal}
\title{Internal GRASP objects}
\alias{.construct_greedy}
\alias{.random_start}
\alias{.objective_function}
\alias{.opt_move_schlecht}
\alias{.k_opt_move}
\alias{.three_opt_move}
\alias{.tsp_from_matrix}
\description{Internal GRASP objects.}
\details{These are not to be called by the user.}
\keyword{internal}