% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack problem solverwith greedy algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x:}{data frame of weights (first column) and value (second column)}

\item{W:}{Maximum capacity (numeric)}
}
\value{
Return the maximum value and elements (index of values) and weight
}
\description{
take the weights, values as parameters and return maximum values and their possible combination
}
