\name{optimal.route}
\alias{optimal.route}
\title{Print the city names of the optimal route.}
\description{
  The function returns the city names of the optimal route generated by
  the function \code{grasp.solve()}.
}
\usage{
optimal.route(x)
}
\arguments{
  \item{x}{\code{x} is the solution generated by \code{grasp.solve()}.}
}
\value{
The function returns a string containing the city names of the optimal route.
}
\references{coming soon ...}
\author{Martin Gartner, Gerold Koeb, Goran Lovric, Stefan Theussl}
\seealso{See also  \code{help.search(grasp.solve)}.}
\examples{
data(dist)
a <- grasp.solve(dist, 29)
optimal.route(a)

## The function is currently defined as
function (x) 
{
    paste(x$city.names[x$route], collapse = " -> ")
  }
}
\keyword{manip}

