\name{plot.tsproblem}
\alias{plot.tsproblem}
\title{Plot a Travelling-Salesman-Problem}
\description{
  Plot the optimal route of a Travelling-Salesman-Problem generated
by the function \code{grasp.solve()}}
}
\usage{
plot.tsproblem(x, ...)
}
\arguments{
  \item{x}{is the solution generated by the function
    \code{grasp.solve()}. }
  \item{...}{\code{...}}
}
\details{
  Unfortunately the plotting doesn't function very well, it still needs
  to be optimized. The horizonal axes need to be exchanged.
}
\value{
A plot of the optimal route.
}
\references{coming soon ...}
\author{Martin Gartner, Gerold Koeb, Goran Lovric, Stefan Theussl}
\seealso{See also  \code{help.search(grasp.solve)}.}
\examples{
data(dist)
a <- grasp.solve(dist, 29)
plot(a)

## The function is currently defined as
function (x, ...) 
{
    coords <- cmdscale(x$matrix)
    plot(-coords[, c(2, 1)], type = "n")
    text(-coords[, c(2, 1)], x$city.names)
    lines(-coords[x$route, c(2, 1)], col = "lightgray")
    mtext(optimal.route(x), side = 3)
  }
}

\keyword{manip}

