\name{print.tsproblem}
\alias{print.tsproblem}
\title{Print-function for a Travelling-Salesman-Problem.}
\description{
  Print a short summary for the solution of a
  Travelling-Salesman-Problem generated by the function \code{grasp.solve()}}
}

\usage{
print.tsproblem(x, ...)
}
\arguments{
  \item{x}{\code{x} is the solution generated by \code{grasp.solve()}.}
  \item{...}{\code{...}}
}
\value{
A short summary of the solution.
}
\references{coming soon ...}
\author{Martin Gartner, Gerold Koeb, Goran Lovric, Stefan Theussl}
\seealso{See also  \code{help.search(grasp.solve)}.}
\examples{
data(dist)
a <- grasp.solve(dist, 29)
a

## The function is currently defined as
function (x, ...) 
{
    if (!inherits(x, "tsproblem")) 
        stop("'x' must inherit from class \"tsp\"")
    cat("A travelling salesman problem with", x$n.cities, "cities\n")
    cat("Minimum distance found so far:", x$optimum, "\n")
    invisible(x)
  }
}

\keyword{manip}

